RESTRUCTURE: Moved some permissions from the messagesController/router into the authcontroller to simplify the MVC from 'crossing wires' on functionality
Seperated concerns to where authController handles
auth and the homepage, and messagesController
handles adding new messages only

Provided some imports that were missing in scripts
that needed them.

Changed app.js router URL .use() to match
changes to the restructure.
