+++ controllers/messagesController.js
@@ -1,34 +1,9 @@
-const { getAllMessages, insertMessage } = require("../db/queries");
+const { insertMessage } = require("../db/queries");
 
-async function displayMessages(req, res) {
-  try {
-    const messages = await getAllMessages();
-
-    //check if the user is a member
-    const isMember =
-      req.session.user && req.session.user.membership === "member";
-
-    //conditional return to the const
-    // what will be sent to the render response
-    const messagesToRender = messages.map((message) => {
-      if (isMember) {
-        //contains messages.title, messages.message AS content,
-        //messages.timestamp,
-        //users.username
-        return message;
-      }
-
-      return {
-        title: message.title,
-        content: message.content,
-        timestamp: message.timestamp,
-      };
-    });
-
-    res.render("index", { title: "Message Board", messages: messagesToRender });
-  } catch (error) {
-    res.status(500).send("Error fetching messages");
-  }
+function renderNewMsgForm(req, res) {
+  res.render("form", {
+    title: "New Message",
+  });
 }
 
 async function postMessage(req, res) {
@@ -47,10 +22,4 @@ async function postMessage(req, res) {
   }
 }
 
-function renderNewMsgForm(req, res) {
-  res.render("form", {
-    title: "New Message",
-  });
-}
-
-module.exports = { displayMessages, postMessage, renderNewMsgForm };
+module.exports = { postMessage, renderNewMsgForm };
