+++ routes/messagesRouter.js
@@ -1,7 +1,6 @@
 const { Router } = require("express");
 
 const {
-  displayMessages,
   postMessage,
   renderNewMsgForm,
 } = require("../controllers/messagesController");
@@ -9,7 +8,7 @@ const {
 const messagesRouter = Router();
 
 //routing
-messagesRouter.get("/", displayMessages);
+
 messagesRouter.get("/new", renderNewMsgForm);
 messagesRouter.post("/new", postMessage);
